pyproject.toml
[project]
name = "betta-mvp"
version = "1.0.0"
description = "多 Agent 舆情分析系统 (对标 BettaFish)"
requires-python = ">=3.10"
# 这是你在项目中用到的所有核心第三方库
dependencies = [
    # 核心基础组件
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    
    # RAG 与向量检索组件
    "sentence-transformers>=2.5.0",
    "faiss-gpu>=1.8.0", # 注意：服务器上如果没有 GPU，装 cpu 版；如果有 GPU，可以改装 faiss-gpu
    "networkx>=3.2.0",  # GraphRAG 需要
    
    # 数据库与 ORM
    "SQLAlchemy>=2.0.0",
    
    # 搜索工具
    "duckduckgo-search>=5.0.0",
    
    # Web 框架与流式输出
    "Flask>=3.0.0",
    "flask-cors>=4.0.0",
    
    # 未来微调与多模态预留 (按需安装，如果服务器硬盘小可先注释掉)
    # "transformers>=4.38.0",
    # "peft>=0.9.0",
    # "datasets>=2.18.0",
    # "accelerate>=0.27.0"
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"